requirements_test.txt
# Test dependencies for Sherbrooke Waste Collection

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0

# Type checking (optional)
mypy>=1.0.0

# Linting
flake8>=6.0.0
pylint>=2.17.0

# Home Assistant dependencies (for testing)
homeassistant>=2024.0.0
aiohttp>=3.8.0
voluptuous>=0.13.0

# Integration dependencies
icalendar>=5.0.0
recurring-ical-events>=2.0.0

# Image validation for brand assets
pillow>=10.0.0
